fix: deprecate verbose composables
