+++ packages/unhead/src/runtime/composables/index.ts
@@ -1,3 +1,4 @@
-export * from './client'
-export * from './server'
+export * from './useHead'
+export * from './useServerHead'
 export * from './useSeoMeta'
+export * from './deprecations'
