+++ packages/unhead/src/runtime/composables/useServerHead.ts
@@ -1,0 +1,10 @@
+import type {
+  ActiveHeadEntry,
+  Head,
+  HeadEntryOptions,
+} from '@unhead/schema'
+import { useHead } from './useHead'
+
+export function useServerHead<T extends Head>(input: T, options: HeadEntryOptions = {}): ActiveHeadEntry<T> | void {
+  return useHead(input, { ...options, mode: 'server' })
+}
